Major rework done, this method seems to be working better so it might be come a maing branch

+++ src/Genetic_Algorithm/Fitness/FitnessImp.java
@@ -11,6 +11,6 @@ public interface FitnessImp {
      * @param dungeon dungeon map
      * @return score that the map achieved
      */
-    int evaluateDungeon(Dungeon dungeon);
+    double evaluateDungeon(Dungeon dungeon);
 }
